+++ src/main/controllers/HomeController.ts
@@ -1,0 +1,5 @@
+export class HomeController {
+  public get(req: any, res: any): void {
+    res.render('home', req.i18n.getDataByLanguage(req.lng).home);
+  }
+}
