--- src/main/controllers/home.ts
@@ -1,3 +1,0 @@
-export const getHomePage = (req: any, res: any) => {
-  res.render('home', req.i18n.getDataByLanguage(req.lng).home);
-};
