+++ src/main/routes.ts
@@ -1,18 +1,11 @@
 import { Application } from 'express';
 
-import {
-  getSearchOption,
-  postSearchOption,
-  getLocationSearch,
-} from './controllers/search';
-import { getHomePage } from './controllers/home';
-
 export default function(app: Application): void {
 
-  app.get('/', getHomePage);
-  app.get('/search-option', getSearchOption);
-  app.post('/search-option', postSearchOption);
-  app.get('/location-search', getLocationSearch);
+  app.get('/', app.locals.container.cradle.homeController.get);
+  app.get('/search-option', app.locals.container.cradle.searchOptionController.get);
+  app.post('/search-option', app.locals.container.cradle.searchOptionController.post);
+  app.get('/location-search', app.locals.container.cradle.locationSearchController.get);
   app.get('/search-for-location', app.locals.container.cradle.searchResultsController.get);
 
 }
