+++ src/test/unit/test/modules/i18next.ts
@@ -1,5 +1,5 @@
-import { I18next } from '../../../main/modules/i18next';
-import { app } from '../../../main/app';
+import { I18next } from '../../../../main/modules/i18next';
+import { app } from '../../../../main/app';
 
 describe('i18next', () => {
   describe('enableFor', () => {
